refactor: move to src layout and use entry points
