+++ pyproject.toml
@@ -1,7 +1,7 @@
 [project]
 name = "apple-release-dates"
 version = "0.1.0"
-description = "Add your description here"
+description = "Tools to fetch and manage Apple product release dates"
 readme = "README.md"
 requires-python = ">=3.12"
 dependencies = [
@@ -13,3 +13,10 @@ dependencies = [
 [build-system]
 requires = ["setuptools>=61.0", "wheel"]
 build-backend = "setuptools.build_meta"
+
+[tool.setuptools]
+package-dir = {"" = "src"}
+packages = ["apple_release_dates"]
+
+[project.scripts]
+dump-apple = "apple_release_dates.dump:cli"
