+++ src/apple_release_dates/__init__.py
@@ -1,0 +1,3 @@
+"""Package for Apple product release dates data fetching and processing."""
+
+__version__ = "0.1.0"
